ELISA/healthbot-backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
SQLAlchemy
pydantic
openai
PyMuPDF
qrcode
passlib[bcrypt]
